Add basic tox.ini configuration
Adds two test environments for running flake8 and
a dry-run of the hub-stress-test script.

We might want to package this into a real dist later
but for now this minimal setup is probably sufficient.

Also needed to change the mock import to pull from the
standard library rather than the third party `mock` package.