Add test to simulate user activity updates
The new command takes a number of users and a number of
worker threads. Each worker thread takes an equal share of
the users and POSTs an activity to update to its user(s)'
API endpoint. Each worker thread sleeps a random amount
of time between 0 and 1 seconds in between processing
each user.

A thread is run alongside the worker threads that just
retrieves a user from /hub/api/users/{user} in a loop.
The time to complete this call is stored and when the
test terminates it shows the average, min, and max time
to complete this API call.

The existing test has been moved under a new subcommand
named 'stress-test'. The activity test can be run with the
'activity-stress-test' command.